added config

+++ src/main/java/com/talhanation/siegeweapons/entities/TransportCartEntity.java
@@ -13,8 +13,6 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
 
-import java.util.Collections;
-
 public class TransportCartEntity extends AbstractInventoryVehicleEntity {
 
     private static final EntityDataAccessor<Integer> CARGO = SynchedEntityData.defineId(TransportCartEntity.class, EntityDataSerializers.INT);
@@ -59,7 +57,7 @@ public class TransportCartEntity extends AbstractInventoryVehicleEntity {
     }
 
     @Override
-    public float getMaxHealth() {
+    public double getMaxHealth() {
         return 200;//TODO: CONFIG
     }
